update delete command

+++ MAS.DappertStorageTest.Cqrs/Delete/DeleteCommandHandler.cs
@@ -1,5 +1,7 @@
 ﻿namespace MAS.DappertStorageTest.Cqrs
 {
+    using System;
+
     using MAS.DapperStorageTest.Infrastructure;
     using MAS.DappertStorageTest.Cqrs.Infrastructure;
 
@@ -40,6 +42,12 @@
         private int DeleteByFilters(DeleteCommand command)
         {
             var (whereCondition, arguments) = BuildWhereFilter(command.EntityName, command.FilterGroup);
+
+            if (string.IsNullOrEmpty(whereCondition))
+            {
+                throw new ArgumentException("Filter doesn't contains any expression.");
+            }
+
             var sqlQuery = BuildQuery($"DELETE FROM [{command.EntityName}] WHERE {whereCondition}");
             var result = 0;
